// fe/src/app/type.ts
import { ReactElement } from "react";

interface DataType {
    id: string |undefined;
    avatar: string |undefined;
    name: string |undefined;
    major: string |undefined;
    grade: string |undefined;
    sex: string |undefined;
    phone: string |undefined;
    email: string |undefined;
    operation: ReactElement |undefined;
  }


export default DataType